Moved WooPosBanner from common package to products package
